requirements.txt
# ---- PyTorch CPU ONLY  ----
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu

# ---- Web framework ----
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ---- Settings and validation ----
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ---- PDF extraction ----
PyMuPDF==1.24.3
pdfplumber==0.11.0

# ---- Vector search ----
faiss-cpu==1.8.0

# ---- Embeddings ----
sentence-transformers==2.7.0

# ---- Keyword search ----
rank-bm25==0.2.2

# ---- LLM provider ----
groq==0.8.0

# ---- Tokenization ----
tiktoken==0.7.0

# ---- Numerics ----
numpy==1.26.4

# ---- HTTP client ----
httpx==0.27.0
